New plugin for Hotkey Swapping
This plug in is to allow a Hotkey to swap to the next game in the shuffler - This is selected via a dropdown to a list of key combinations using Insert. This will allow various bots and such to call on a swap for things like Twitch channel point redeems.