Resolve extra swaps in MMXTreme 2
Relies on swap_exceptions being added to generic_swap - but possibly could be implemented another way.

When going to Options or Boss Attack, lives are changed. This shouldn't swap.

When lives go from 255 (game over) to 2 or 0 (started playing again, in regular or Boss Attack), this shouldn't swap. Treat 255 like 0.

When X or Zero tags in and has lower HP than their partner, this shouldn't swap.